move selectedCollcocateDeviceSlice to collocation service folder
