+++ platform/src/common/components/Collocation/DeviceStatus/Table/DataTable.jsx
@@ -5,7 +5,7 @@ import {
   addDevices,
   removeDevices,
   addDevice,
-} from '@/lib/store/services/addMonitor/selectedCollocateDevicesSlice';
+} from '@/lib/store/services/collocation/selectedCollocateDevicesSlice';
 import Skeleton from './Skeleton';
 
 const DataTable = ({ filteredData, collocationDevices, isLoading }) => {
