+++ platform/src/lib/store/services/collocation/selectedCollocateDevicesSlice.js
